tests: split test_utils.py into per-module utils test files
One test_*.py per utils/*.py, matching the regression/profiling 1-to-1
naming convention: test_astro.py, test_kpf_utils.py, test_stats.py,
test_config.py. Pure move of the existing classes -- no test logic changed
(86 tests, unchanged). io.py and logger.py already have their coverage in
test_masters_recipe.py and test_logger.py respectively.
